Implementation cleanup
