+++ src/Microsoft.Diagnostics.Runtime/src/Implementation/ClrExtendedModuleData.cs
@@ -1,0 +1,15 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace Microsoft.Diagnostics.Runtime.Implementation
+{
+    internal class ClrExtendedModuleData
+    {
+        public string? SimpleName { get; set; }
+        public string? FileName { get; set; }
+        public bool IsFlatLayout { get; set; }
+        public bool IsDynamic { get; set; }
+        public ulong Size { get; set; }
+    }
+}
