+++ src/Microsoft.Diagnostics.Runtime/src/Implementation/ClrPrimitiveType.cs
@@ -66,6 +66,7 @@ namespace Microsoft.Diagnostics.Runtime.Implementation
         };
 
         public override ulong GetArrayElementAddress(ulong objRef, int index) => 0;
+
         public override T[]? ReadArrayElements<T>(ulong objRef, int start, int count) => null;
 
         public override ClrInstanceField? GetFieldByName(string name) => null;
