+++ src/Microsoft.Diagnostics.Runtime/src/Implementation/ClrRuntimeHelpers.cs
@@ -8,39 +8,6 @@ using System.Threading;
 
 namespace Microsoft.Diagnostics.Runtime.Implementation
 {
-    internal interface IClrRuntimeHelpers : IDisposable
-    {
-        void Flush();
-        IEnumerable<ClrThread> EnumerateThreads();
-        ClrHeap CreateHeap();
-        ClrAppDomainData GetAppDomainData();
-        ClrMethod? GetMethodByMethodDesc(ulong methodDesc);
-        ClrMethod? GetMethodByInstructionPointer(ulong ip);
-        IEnumerable<ClrHandle> EnumerateHandles();
-        IEnumerable<ClrJitManager> EnumerateClrJitManagers();
-        string? GetJitHelperFunctionName(ulong address);
-    }
-
-    internal class ClrAppDomainData
-    {
-        public ClrAppDomain? SystemDomain { get; set; }
-        public ClrAppDomain? SharedDomain { get; set; }
-        public ImmutableArray<ClrAppDomain> AppDomains { get; set; }
-        public Dictionary<ulong, ClrModule> Modules { get; } = new();
-        public ClrModule? BaseClassLibrary { get; set; }
-
-        internal ClrAppDomain? GetDomainByAddress(ulong address)
-        {
-            if (SystemDomain is not null && SystemDomain.Address == address)
-                return SystemDomain;
-
-            if (SharedDomain is not null && SharedDomain.Address == address)
-                return SharedDomain;
-
-            return AppDomains.FirstOrDefault(x => x.Address == address);
-        }
-    }
-
     internal unsafe class ClrRuntimeHelpers : IClrRuntimeHelpers, IClrAppDomainHelpers
     {
         private ClrRuntime? _runtime;
