+++ src/Microsoft.Diagnostics.Runtime/src/Implementation/IClrRuntimeHelpers.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace Microsoft.Diagnostics.Runtime.Implementation
+{
+    internal interface IClrRuntimeHelpers : IDisposable
+    {
+        void Flush();
+        IEnumerable<ClrThread> EnumerateThreads();
+        ClrHeap CreateHeap();
+        ClrAppDomainData GetAppDomainData();
+        ClrMethod? GetMethodByMethodDesc(ulong methodDesc);
+        ClrMethod? GetMethodByInstructionPointer(ulong ip);
+        IEnumerable<ClrHandle> EnumerateHandles();
+        IEnumerable<ClrJitManager> EnumerateClrJitManagers();
+        string? GetJitHelperFunctionName(ulong address);
+    }
+}
