+++ src/Microsoft.Diagnostics.Runtime/src/Implementation/IJitManagerHelpers.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Microsoft.Diagnostics.Runtime.Implementation
+{
+    // Helpers to evaluate lazy fields.
+    public interface IJitManagerHelpers
+    {
+        IEnumerable<ClrNativeHeapInfo> EnumerateNativeHeaps(ClrJitManager clrJitManager);
+    }
+}
